Fix reverse marshalling of refs to blittable types (#8140) (#36119)
Hit in selfhosted ilc.exe. In JitInterface we have a couple cases where we do reverse p/invoke with a delegate that has a ref parameter. Sometimes the `ref` is null (and we don't touch it).

The marshaller was trying to make a copy of the pointed to value, NullRefing in the process because the pointer was null. We don't actually need to make a copy because the value is blittable.

Co-authored-by: Michal Strehovský <MichalStrehovsky@users.noreply.github.com>