new GetGCMemoryInfo API (#37879)
* new GetGCMemoryInfo API

This implements the new GetGCMemoryInfo API that provides much more info than the one we shipped in 3.0. Also added an overload to query GCs by kind - ephemeral, full blocking or BGC, to facilitate not having to keep a really small sampling interval but can still get the kinds of GCs that happen very infrequently.

Most of the changes inside GC are the following -

All the info is updated at the end of a GC with the exception of BGC which needs special care. We maintain 2 BGC info data structures and one could be in the process of getting updated and if the user requests either BGC info or the last GC info which is also a BGC, we would return the one that's complete. See comments in code for a more detailed explanation.

Fill in the after size data for all generations instead of just the affected generations.

Timing info recording -

+ Changed the GC time recording from ms to us. Previously we did the timing analysis with events which also carry their own high res timestamps so in GC using ms is good enough. But now since we are returning the timing info we want to record it in us instead of ms.
+ Consolidated when we get end_gc_time. It was not including all of the GC work before (eg, decommit eph seg). And moving it to do_post_gc makes it easier for getting the GC info.
+ Not including the EE restart time for blocking GCs - the reason being restart takes very little time and by the time we get to the actual RestartEE the GC settings could be changed and it's not worth the effort to wait till RestartEE at which point we have to spend more effort to figure out which GC to attribute it to.
+ Maintain the running % Pause time in GC info.