Add telemetry to System.Net.Http (#37619)
* Add telemetry to System.Net.Http

This is the first of several PRs to add telemetry focused events and counters.

A few things to note:

* Added a new dedicated namespace for telemetry. Current feedback from partners is that our
existing events are "confusing and a mess" due to their focus on low-level diagnostics and verbose
debugging. See discussion in #37428 regarding reasons for a new namespace.

* Please comment on the naming patterns for the events, counters, and
display names, etc. I looked at ASP.NET events/counters and @anurse older guidance in
https://gist.github.com/anurse/af1859663ac91c6cf69c820cebe92303. But there are inconsistencies
between all of that. So, I'd like to get agreement on naming patterns for events/counters
(i.e. noun-verb vs. verb-noun, present tense vs. past tense, etc.).

* No automated tests were added at this time. The current discussions with others (reverse proxy
team) is that automated tests are brittle in CI. But manual tests will be run.

Contributes to #37428

* Address PR feedback

* PR feedback
* Added RequestAbort as separate event method with dimensions (host, port)

* Address PR feedback 2

* RequestStart now includes host, port, scheme, path and version
* RequestStart and RequestStop now bracket the lifetime of the request including reading the
response body stream. If the response body is not complete read then disposing the stream
will trigger RequestStop as well.

* Address feedback 3

* Use exception filter to improve performance

* Avoid negative 'current-requests' counter value

* Address feedback 4