Ignore 0-byte responses from AIA fetch.
When fetching a certificate with AIA, we should ignore responses that
are zero bytes in length. A zero-byte value passed to X509Certificate2
will create a default certificate with a null PAL. This was being
passed in to the OpenSSL chain builder, and would pass in a null handle
for the certificate.