Delete APM fallback from Socket task operations (#39248)
* Fix static field names

* Remove CachedEventArgs wrapper

This was put in place initially when the Task-based APIs were new, and we wanted to make them pay-for-play such that other uses of Socket wouldn't incur the extra fields.  But at this point, the Task-based APIs are the primary/recommended way to use Socket, and when they are used, the wrapper adds overhead (both in size and CPU).  Deleting it.

* Delete APM fallback from AcceptAsync

* Delete APM fallbacks for Receive/SendAsync with tasks