Use unbounded connected streams in System.Net.Security tests (#43327)
Recently there's been a rash of sporadic timeouts/hangs in System.Net.Security tests.  I don't _think_ it's related to hitting the maximum buffer size in ConnectedStreams (e.g. if the test is written in a way that it assumes a write will always complete synchronously and won't proceed to read until it does), but I'm removing the bound to see if that helps.  We can put it back later if desired: it's not what the tests are validating, and previous to my ConnectedStream change there was no bound, either.