Don't batch the SuperPMI collection job (#45861)
The SuperPMI collection job gets triggered whenever the JIT-EE GUID changes.
We need a collection every time that happens. If, in the unlikely circumstance,
two GUID changes happen in quick succession, we don't want to skip
one due to batching, as that would leave us a window where there are
no collections corresponding to a range of commits.